requirements.txt
requests
beautifulsoup4
python-dotenv
duckduckgo-search
ollama
flask
openai
gunicorn
